requirements.txt
pytest==7.*
stomp.py==8.1.* 
pysiddhi